refactor: package 분리, view 분리
