--- src/main/java/racingcar/RandomUtil.java
@@ -1,12 +1,0 @@
-package racingcar;
-
-import java.util.Random;
-
-public class RandomUtil {
-    static final int BOUND = 10;
-
-    public int getRandomInt() {
-        Random random = new Random();
-        return random.nextInt(BOUND);
-    }
-}
