+++ src/main/java/racingcar/model/Car.java
@@ -1,0 +1,17 @@
+package racingcar.model;
+
+class Car {
+    private int score;
+
+    Car() {
+        score = 0;
+    }
+
+    void go() {
+        score++;
+    }
+
+    int getScore() {
+        return this.score;
+    }
+}
