+++ src/main/java/racingcar/service/RandomService.java
@@ -1,0 +1,12 @@
+package racingcar.service;
+
+import java.util.Random;
+
+public class RandomService {
+    private static final int BOUND = 10;
+
+    public int getRandomInt() {
+        Random random = new java.util.Random();
+        return random.nextInt(BOUND);
+    }
+}
