+++ src/main/java/racingcar/view/InputView.java
@@ -1,11 +1,10 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.Scanner;
 
-public class GuideView {
+public class InputView {
     private static final String MSG_CARS="자동차 대수는 몇 대 인가요?";
     private static final String MSG_ATTEMPTS="시도할 회수는 몇 회 인가요?";
-    private static final String MSG_RESULT="실행 결과";
 
     public int inputNumberOfCars() {
         System.out.println(MSG_CARS);
@@ -18,9 +17,4 @@ public class GuideView {
         Scanner scanner = new Scanner(System.in);
         return scanner.nextInt();
     }
-
-    public void printResultMessage(){
-        System.out.println(MSG_RESULT);
-    }
-
 }
