+++ src/test/java/racingcar/model/CarsTest.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.Cars;
 
 import java.util.List;
